# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "iOS Screenshot Processing API",
    "description": "Multi-user platform for processing iPhone screen time and battery screenshots",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/consensus/groups": {
      "get": {
        "tags": [
          "Consensus"
        ],
        "summary": "Get Groups With Verification Tiers",
        "description": "Get all groups with verification tier breakdown.\n\nReturns groups with counts for:\n- single_verified: Verified by exactly 1 user\n- agreed: 2+ users verified, all values match exactly\n- disputed: 2+ users verified, any difference exists",
        "operationId": "get_groups_with_verification_tiers_api_v1_consensus_groups_get",
        "parameters": [
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupVerificationSummary"
                  },
                  "title": "Response Get Groups With Verification Tiers Api V1 Consensus Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/groups/{group_id}/screenshots": {
      "get": {
        "tags": [
          "Consensus"
        ],
        "summary": "Get Screenshots By Tier",
        "description": "Get screenshots in a specific verification tier for a group.",
        "operationId": "get_screenshots_by_tier_api_v1_consensus_groups__group_id__screenshots_get",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "single_verified",
                "agreed",
                "disputed"
              ],
              "type": "string",
              "description": "Verification tier to filter by",
              "title": "Tier"
            },
            "description": "Verification tier to filter by"
          },
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreenshotTierItem"
                  },
                  "title": "Response Get Screenshots By Tier Api V1 Consensus Groups  Group Id  Screenshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/screenshots/{screenshot_id}/compare": {
      "get": {
        "tags": [
          "Consensus"
        ],
        "summary": "Get Screenshot Comparison",
        "description": "Get comparison data for a screenshot with multiple verifiers.\nShows each verifier's annotation and highlights differences.",
        "operationId": "get_screenshot_comparison_api_v1_consensus_screenshots__screenshot_id__compare_get",
        "parameters": [
          {
            "name": "screenshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screenshot Id"
            }
          },
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenshotComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/screenshots/{screenshot_id}/resolve": {
      "post": {
        "tags": [
          "Consensus"
        ],
        "summary": "Resolve Dispute",
        "description": "Resolve a dispute by setting the final agreed-upon values.\n\nRequires admin role. Original annotations are preserved - resolved values\nare stored separately on the screenshot.",
        "operationId": "resolve_dispute_api_v1_consensus_screenshots__screenshot_id__resolve_post",
        "parameters": [
          {
            "name": "screenshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screenshot Id"
            }
          },
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveDisputeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/summary/stats": {
      "get": {
        "tags": [
          "Consensus"
        ],
        "summary": "Get Consensus Summary",
        "description": "Get summary statistics for consensus analysis.",
        "operationId": "get_consensus_summary_api_v1_consensus_summary_stats_get",
        "parameters": [
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsensusSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/{screenshot_id}": {
      "get": {
        "tags": [
          "Consensus"
        ],
        "summary": "Get Consensus Analysis",
        "description": "Get consensus analysis for a specific screenshot (legacy endpoint).",
        "operationId": "get_consensus_analysis_api_v1_consensus__screenshot_id__get",
        "parameters": [
          {
            "name": "screenshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screenshot Id"
            }
          },
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsensusAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consensus/{screenshot_id}/recalculate": {
      "post": {
        "tags": [
          "Consensus"
        ],
        "summary": "Recalculate Consensus",
        "description": "Recalculate consensus for a specific screenshot (legacy endpoint).",
        "operationId": "recalculate_consensus_api_v1_consensus__screenshot_id__recalculate_post",
        "parameters": [
          {
            "name": "screenshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Screenshot Id"
            }
          },
          {
            "name": "x-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsensusAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupVerificationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image_type": {
            "type": "string",
            "enum": [
              "battery",
              "screen_time"
            ],
            "title": "Image Type"
          },
          "single_verified": {
            "type": "integer",
            "title": "Single Verified",
            "default": 0
          },
          "agreed": {
            "type": "integer",
            "title": "Agreed",
            "default": 0
          },
          "disputed": {
            "type": "integer",
            "title": "Disputed",
            "default": 0
          },
          "total_verified": {
            "type": "integer",
            "title": "Total Verified",
            "default": 0
          },
          "total_screenshots": {
            "type": "integer",
            "title": "Total Screenshots",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "image_type"
        ],
        "title": "GroupVerificationSummary",
        "description": "Group with verification tier breakdown."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveDisputeRequest": {
        "properties": {
          "hourly_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Hourly Values"
          },
          "extracted_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Title"
          },
          "extracted_total": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Total"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          }
        },
        "type": "object",
        "required": [
          "hourly_values"
        ],
        "title": "ResolveDisputeRequest",
        "description": "Request to resolve a dispute with final values."
      },
...
      "ScreenshotTierItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "participant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Id"
          },
          "screenshot_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot Date"
          },
          "verifier_count": {
            "type": "integer",
            "title": "Verifier Count"
          },
          "has_differences": {
            "type": "boolean",
            "title": "Has Differences"
          },
          "extracted_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_path",
          "participant_id",
          "screenshot_date",
          "verifier_count",
          "has_differences"
        ],
        "title": "ScreenshotTierItem",
        "description": "Screenshot summary for tier list view."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}